Improved inventory recognition HUD for pad
Inventory HUD when using pad is rather difficult to properly
recognize since the pg info moves when user move the left
analog joystick.
With this change the gray line used to identify the pg info
is searched within an area with a basic algorithm that searches
for two horizontal gray lines at the correct distance.